More efficient parameterized unit tests
Since the project needs a lot of unit tests, it can benefit from a more
compact and efficient way to write them.

Using "parameterized" unit tests you can define the test once and just
write a lot of parameter sets.
See the improvement in `TestCanonicalDomain` with this PR.
I suggest to use this everywhere.

Unfortunately, python's stdlib `unittest' doesn't support it and we need
to add dependency `parameterized` in `tests_require`. (`pytest` supports
this out of the box).