Use latest htmldate and pass datetime max_date instead of string
The latest `htmldate` is more efficient https://github.com/adbar/htmldate/blob/master/CHANGELOG.md
Update to the latest.

`max_date` can be either a `str` or a `datetime` object.
Since we already have a `datetime` object, there is no point in making
it a string to pass it to `htmldate_find_date`